change to dropdown menu for map hazards
